requirements.txt
pytorch>=1.8
torchvision
opencv-python
